+++ app/decorators/category_decorator.rb
@@ -1,0 +1,10 @@
+class CategoryDecorator < ApplicationDecorator
+  delegate_all
+
+  def case_form_json
+    {
+      id: id,
+      name: name,
+    }
+  end
+end
