+++ app/decorators/service_decorator.rb
@@ -38,7 +38,7 @@ class ServiceDecorator < ClusterPartDecorator
       .transform_keys do |category|
       category.nil? ? Category.new(name: 'Other', id: -1) : category
     end.map do |category, issues|
-      category.case_form_json.merge(issues: issues.map(&:case_form_json))
+      category.decorate.case_form_json.merge(issues: issues.map(&:case_form_json))
     end.reject(&:nil?)
   end
