+++ app/models/category.rb
@@ -5,13 +5,6 @@ class Category < ApplicationRecord
 
   validates :name, presence: true
 
-  def case_form_json
-    {
-      id: id,
-      name: name,
-    }
-  end
-
   def self.globally_available?
     true
   end
